requirements.txt
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
Flask==3.1.0
flatbuffers==25.1.24
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.20.1
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
PyMatting==1.1.13
referencing==0.36.2
rembg==2.0.62
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.1
scipy==1.15.1
sympy==1.13.3
tifffile==2025.1.10
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
gunicorn==23.0.0

